src/requirements.txt
Flask==1.1.1
Flask-Cors==3.0.8
nltk==3.4.5
requests==2.21.0
sklearn==0.0
pandas==0.25.3
pymongo==3.10.1
python-dotenv==0.10.5
dnspython





